CLICKHOUSE-905: reproduce hanging alters on slow inserts
test suite is not a test per se. It allows to run a multiple concurrent threads
on the emulated slow network and also spawn a few dedicated slow running inserts.

the following scenario is affecting merges/inserts to hang on 20.3 but not 20.8.

list of events sorted accordingly of the occurrence

1. Slow insert to table t (long running inserts with a very low throughput)
2. Alter on the table t to change a column type
3. Following alters, inserts and merges stuck on a the lock from the step (2)